VLM_Evaluation_Workspace/requirements.txt
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
transformers==4.44.2
accelerate>=1.0.0
bitsandbytes>=0.43.0
Pillow>=10.0.0
nltk>=3.8
rouge-score>=0.1.2
sentencepiece>=0.2.0
protobuf>=4.0
einops>=0.7.0
timm>=0.9.0
hf_transfer
tiktoken
transformers_stream_generator==0.0.5
matplotlib
